fix(DB-365): 크루 가입 로직 수정
