+++ dongsan-external-api/src/main/java/com/dongsan/api/domains/crew/CrewInfoFacade.java
@@ -14,14 +14,11 @@ import com.dongsan.domain.domains.member.Member;
 import com.dongsan.domain.domains.member.MemberRdbService;
 import com.dongsan.domain.domains.walkwayLog.WalkwayLog;
 import com.dongsan.domain.domains.walkwayLog.WalkwayLogRdbService;
-import com.dongsan.domain.lock.RedisLockService;
 import com.dongsan.domain.support.error.CoreErrorCode;
 import com.dongsan.domain.support.error.CoreException;
 import com.dongsan.domain.support.paging.CursorRequest;
 import com.dongsan.domain.support.paging.CursorResponse;
 import com.dongsan.file.service.S3FileService;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import org.springframework.web.multipart.MultipartFile;
@@ -32,26 +29,24 @@ import java.util.Map;
 
 @Service
 public class CrewInfoFacade {
-    private static final Logger log = LoggerFactory.getLogger(CrewInfoFacade.class);
-
     private final CrewRdbService crewRdbService;
     private final CrewMemberRdbService crewMemberRdbService;
     private final WalkwayLogRdbService walkwayLogRdbService;
     private final MemberRdbService memberRdbService;
     private final S3FileService s3FileService;
     private final ImageRdbService imageRdbService;
-    private final RedisLockService redisLockService;
+    private final CrewJoinFacade crewJoinFacade;
 
     public CrewInfoFacade(CrewRdbService crewRdbService, CrewMemberRdbService crewMemberRdbService,
                           WalkwayLogRdbService walkwayLogRdbService, MemberRdbService memberRdbService, S3FileService s3FileService,
-                          ImageRdbService imageRdbService, RedisLockService redisLockService) {
+                          ImageRdbService imageRdbService, CrewJoinFacade crewJoinFacade) {
         this.crewRdbService = crewRdbService;
         this.crewMemberRdbService = crewMemberRdbService;
         this.walkwayLogRdbService = walkwayLogRdbService;
         this.memberRdbService = memberRdbService;
         this.s3FileService = s3FileService;
         this.imageRdbService = imageRdbService;
-        this.redisLockService = redisLockService;
+        this.crewJoinFacade = crewJoinFacade;
     }
 
     @Transactional(readOnly = true)
@@ -154,31 +149,10 @@ public class CrewInfoFacade {
     public void joinCrew(Long crewId, Long memberId, String password) {
         Crew crew = crewRdbService.getCrew(crewId);
         if (crew.isLimitedCrew()) {
-            joinLimitedCrew(crew, memberId, password);
+            crewJoinFacade.joinLimitedCrew(crew, memberId, password);
         } else {
-            joinUnLimitedCrew(crew, memberId, password);
+            crewJoinFacade.joinUnLimitedCrew(crew, memberId, password);
         }
-    }
-
-    @Transactional
-    public void joinUnLimitedCrew(Crew crew, Long memberId, String password) {
-        crewMemberRdbService.validateNotAlreadyJoined(crew.getId(), memberId);
-        crewRdbService.comparePassword(crew, password);
-        crewMemberRdbService.joinCrew(crew.getId(), memberId);
-    }
-
-    private void joinLimitedCrew(Crew crew, Long memberId, String password) {
-        redisLockService.callWithLock(crew.getId(), () -> {
-            crewRdbService.comparePassword(crew, password);
-
-            int memberCount = crewMemberRdbService.countCrewMember(crew.getId());
-            log.info("memberCount : {}", memberCount);
-            crew.validateNotFull(memberCount);
-            crewMemberRdbService.validateNotAlreadyJoined(crew.getId(), memberId);
-            crewMemberRdbService.joinCrew(crew.getId(), memberId);
-
-            return null;
-        });
     }
 
     public CursorResponse<GetCrewsResponse> getMyCrews(Long memberId, CursorRequest paging) {
